requirements.txt
altair
pandas
streamlit
Pillow
python-dotenv
google-generativeai
